[backend] fix: DiaryServiceImpl.updateDiary에서 added/removedTeamIds가 비어있을 경우 팀 연동 쿼리 실행하지 않도록 수정

+++ backend/src/main/java/com/example/demo/service/DiaryServiceImpl.java
@@ -42,8 +42,13 @@ public class DiaryServiceImpl implements DiaryService {
         diaryRepository.updateDiaryV2(diaryEditRequestDTOv2);
 
         Long diaryId = diaryEditRequestDTOv2.getId();
-        teamDiaryRepository.insertTeamDiaryV2(diaryId, diaryEditRequestDTOv2.getAddedTeamIds());
-        teamDiaryRepository.deleteTeamDiaryV2(diaryId, diaryEditRequestDTOv2.getRemovedTeamIds());
+        if (!diaryEditRequestDTOv2.getAddedTeamIds().isEmpty()) {
+            teamDiaryRepository.insertTeamDiaryV2(diaryId, diaryEditRequestDTOv2.getAddedTeamIds());
+        }
+
+        if (!diaryEditRequestDTOv2.getRemovedTeamIds().isEmpty()) {
+            teamDiaryRepository.deleteTeamDiaryV2(diaryId, diaryEditRequestDTOv2.getRemovedTeamIds());
+        }
     }
 
     // 선택한 다이어리 상세 정보 요청
